Gradually fade PTT output between 0 and high impedance.
This is in order to reduce audible noise coupling from the PTT line
into the transmitted audio due to a potential schematic/PCB layout
issue with v0.7. With this change, thumping noises are no longer
audible on the right channel of the radio jack when in HW test mode
(especially when in a mode that silences the radio jack).

Note that this should *really* be fixed in hardware as the extra
noise can still be recorded by programs like Audacity using a good
enough sound device. This also reduces the effective PWM resolution,
although it's very difficult for a human to notice a difference
in LED brightness at the former resolution. TBD for later.