v0.8 board: Increase TP4056 CE capacitor value from 1uF to 10uF.
This is due to the USB inrush test failing (TP4056 turns on after 15ms, not >100ms as required).
While 1uF should have been sufficient combined with a 100K resistor in theory, DC bias properties
for SMD X5R/X7R capacitors resulted in losing ~75% of the rated capacitance on 0402 sized parts.
This calculates to an actual RC time constant of ~0.023s instead of >= 0.1s.

(Note: in limited testing, I do not seem to be seeing any overcurrent errors or other weirdness
when plugging ezDV into USB hosts/power supplies. This may not be a huge issue in practice, but
should still be fixed in the interest of ensuring a good user experience.)