Improve algorithm to determine the number of packets to send.
The Flex code now keeps a running total of how much extra time is required
in between timer ticks. It then calculates the number of extra VITA packets
required to be sent based on how many us we're behind what the radio expects,
ensuring that the radio receives on average the correct number of packets
in a given time interval.